move abc to interface file
